pyproject.toml
[tool.poetry]
name = "ame-team"
version = "0.1.0"
description = ""
authors = ["armanisadeghi <132974515+armanisadeghi@users.noreply.github.com>"]
readme = "README.md"

packages = [{include = "ame-team"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
python-decouple = "^3.8"
openai = "^1.26.0"
bs4 = "^0.0.2"
scrapy = "^2.11.1"
selenium = "^4.20.0"
markdown-it-py = "^3.0.0"
aiofiles = "^23.2.1"
mdit-py-plugins = "^0.4.0"
asgiref = "^3.8.1"
yt-dlp = "^2024.4.9"
moviepy = "^1.0.3"
ffmpeg = "^1.4"
aiohttp = "^3.9.5"
aiomultiprocess = "^0.9.1"
pymupdf = "^1.24.2"
pytesseract = "^0.3.10"
tabula = "^1.0.5"
python-pptx = "^0.6.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
